refactor: rewrite loading indicator for search results: put it into store, avoid prop-drilling
