+++ src/components/search-results/search-results.tsx
@@ -6,10 +6,11 @@ import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import Pagination from '../pagination/pagination';
 import PersonInList from '@components/person-in-list/person-in-list';
 import { IPerson } from '@services/api-types';
+import { useSelector } from 'react-redux';
+import { isPeopleLoadingSelector } from '../../store';
 
 export interface SearchResultsProps extends ComponentProps<'div'> {
   searchResults?: TPeopleReponse;
-  isLoading: boolean;
   setPageCallback: (value: number) => void;
 }
 
@@ -40,8 +41,9 @@ export default function SearchResults(props: SearchResultsProps) {
   const navigate = useNavigate();
   const params = useParams();
   const location = useLocation();
+  const isLoading = useSelector(isPeopleLoadingSelector).value;
 
-  if (props.isLoading) {
+  if (isLoading) {
     return (
       <div className={[styles.searchResults, styles.resultsFrame].join(' ')}>
         <Loader />
